+++ auto-matching-api-rails/lib/auto_matching/common/pcmax.rb
@@ -1,7 +1,9 @@
 module AutoMatching
-  module Checker
-    class Pcmax < CheckerBase
-      include Common::Pcmax
+  module Common
+    module Pcmax
+      def source_site_key
+        SourceSite::KEY_PCMAX
+      end
 
       def try_login
         session.first(".register-header > .register-h2#login-tab").click
