+++ auto-matching-api-rails/lib/auto_matching/reader/ikukuru.rb
@@ -3,9 +3,10 @@ module AutoMatching
     class Ikukuru < CheckerBase
       include Common::Ikukuru
 
-      def read_board
-        # TODO
-      end
+      private
+        def read_board
+          # TODO
+        end
     end
   end
 end
