+++ auto-matching-api-rails/lib/auto_matching/reader/merupara.rb
@@ -3,6 +3,8 @@ module AutoMatching
     class Merupara < CheckerBase
       include Common::Merupara
     end
+
+    private
     def read_board
       # TODO
     end
