+++ auto-matching-api-rails/lib/auto_matching/sender/merupara.rb
@@ -1,19 +1,9 @@
 module AutoMatching
   module Sender
     class Merupara < SenderBase
-      class << self
-        def source_site_key
-          SourceSite::KEY_MERUPARA
-        end
-      end
+      include Common::Merupara
 
       private
-        def try_login
-          session.fill_in "loginid", with: login_user
-          session.fill_in "pwd", with: login_password
-          session.find("#B1login").click
-        end
-
         def delete_past_post
           session.visit("https://meru-para.com/ms/mb/MB.aspx?sid=&pf=7")
