+++ auto-matching-api-rails/lib/auto_matching/sender/mint.rb
@@ -1,20 +1,9 @@
 module AutoMatching
   module Sender
     class Mint < SenderBase
-      class << self
-        def source_site_key
-          SourceSite::KEY_MINT
-        end
-      end
+      include Common::Mint
 
       private
-
-        def try_login
-          session.fill_in "loginid", with: login_user
-          session.fill_in "pwd", with: login_password
-          session.find("#B1login").click
-        end
-
         def delete_past_post
           session.visit("https://mintj.com/ms/mb/MB.aspx?sid=&pf=7")
           session.click_link("③投稿一覧")
