+++ auto-matching-api-rails/lib/auto_matching/sender/wakuwaku.rb
@@ -3,20 +3,9 @@
 module AutoMatching
   module Sender
     class Wakuwaku < SenderBase
-      class << self
-        def source_site_key
-          SourceSite::KEY_WAKUWAKU
-        end
-      end
+      include Common::Wakuwaku
 
       private
-
-        def try_login
-          session.fill_in "email", with: login_user
-          session.fill_in "password", with: login_password
-          session.click_on "ログイン"
-        end
-
         def delete_past_post
           session.visit("http://550909.com/m/bbs/history")
